ACL is now minimally working
This depends on pull request #20 of Ethernet2
(https://github.com/adafruit/Ethernet2/pull/20). In the meanwhile,
patching the library with the commit works well.

This makes us able to 403 to unknown IP addresses not included in the
ACL. We do not reply to authorized clients yet, but code-wise it seems
to work.